requirements.txt
jupyter
matplotlib
numpy
pillow
pyperclip
tensorflow
types-tensorflow
-e /Users/pictomo/Repositories/HumanAI.io